# doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000" #Api runs locally in port 3000
info:
  title: WASAPhoto
  description: "WASAPhoto API documentation"
  version: 1.0.0
components:
  securitySchemes:  #Security of API, for authentication use Bearer where token represented in JWT Format
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /users:     #2)path for retrieving users
    get:      # using get since we are retrieving information rather than modifying or creating anything.
      tags:
        - Users
      summary: "Retrieves a list of users"
      description: "Fetches users based on a search query."
      operationId: getUsers
      security:
        - BearerAuth: [] #request requires authentication.
      parameters:
        - name: search
          in: query
          description: "Search query to filter users"
          required: true
          schema:
            type: string
            example: "Maria"
            minLength: 1
            maxLength: 100
            pattern: ".*"
      responses:
        '200':
          description: "List of matching users"
          content:
            application/json:
              schema:
                type: object
                description: "Response containing a list of users"
                properties:
                  users:
                    type: array
                    description: "List of user objects"
                    minItems: 0
                    maxItems: 1000
                    items:
                      type: object
                      description: "A user object"
                      properties:
                        id:
                          type: integer
                          format: int64
                          description: "User ID"
                          example: 123
                        username:
                          type: string
                          description: "Username"
                          example: "Maria"
                          minLength: 3
                          maxLength: 32
                          pattern: "^[a-zA-Z0-9_]+$"
                        profilePhotoUrl:
                          type: string
                          description: "URL of the user's profile photo"
                          example: "https://example.com/profile.jpg"
                          minLength: 5
                          maxLength: 256
                          pattern: "^(https?://.+)$"
                        createdAt:
                          type: string
                          format: date-time
                          description: "Timestamp when the user was created"
                          example: "2025-02-11T12:34:56Z"
                          minLength: 20
                          maxLength: 30
                          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"
                        updatedAt:
                          type: string
                          format: date-time
                          description: "Timestamp when the user was last updated"
                          example: "2025-02-11T14:56:22Z"
                          minLength: 20
                          maxLength: 30
                          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"
                      required:
                        - id
                        - username
                        - profilePhotoUrl
                        - createdAt
                        - updatedAt
                required:
                  - users
        '400':
          description: "Bad request (missing search query)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Query is empty"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
...
  /message/send:
    post:
      tags:
        - Messages
      summary: "Send a message"
      description: "Allows a user to send a message to another user or a group."
      operationId: sendMessage
      security:
        - BearerAuth: []
      requestBody:
        description: "Message sending request"
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Request object for sending a message"
              properties:
                recipientId:
                  type: integer
                  format: int64
                  description: "ID of the recipient (user or group)"
                  example: 456
                isGroup:
                  type: boolean
                  description: "Specifies whether the recipient is a group"
                  example: false
                content:
                  type: string
                  description: "The message content"
                  example: "Hello, how are you?"
                  minLength: 1
                  maxLength: 1000
                  pattern: "^[\\s\\S]+$"
              required:
                - recipientId
                - isGroup
                - content
      responses:
        '200':
          description: "Message sent successfully"
          content:
            application/json:
              schema:
                type: object
                description: "Response containing details of the sent message"
                properties:
                  messageId:
                    type: integer
                    format: int64
                    description: "ID of the sent message"
                    example: 789
                  senderId:
                    type: integer
                    format: int64
                    description: "ID of the sender"
                    example: 123
                  recipientId:
                    type: integer
                    format: int64
                    description: "ID of the recipient"
                    example: 456
                  content:
                    type: string
                    description: "The message content"
                    example: "Hello, how are you?"
                    minLength: 1
                    maxLength: 1000
                    pattern: "^[\\s\\S]+$"
                  timestamp:
                    type: string
                    format: date-time
                    description: "Timestamp of when the message was sent"
                    example: "2025-02-11T15:00:00Z"
                    minLength: 20
                    maxLength: 30
                    pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"
                required:
                  - messageId
                  - senderId
                  - recipientId
                  - content
                  - timestamp
        '400':
          description: "Bad request (invalid request payload)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Failed to send message"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
        '401':
          description: "Unauthorized (user not authenticated)"
          content:
            application/json:
              schema:
                type: object
                description: "Unauthorized error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Unauthorized access"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
...
  /messages:
    post:
      tags:
        - Messages
      summary: "Retrieve messages"
      description: "Fetches messages for a specific user or conversation."
      operationId: getConversation
      security:
        - BearerAuth: []
      requestBody:
        description: "Request to retrieve messages"
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Request object for retrieving messages"
              properties:
                conversationId:
                  type: integer
                  format: int64
                  description: "ID of the conversation"
                  example: 123
                recipientId:
                  type: integer
                  format: int64
                  description: "ID of the user or group to fetch messages from"
                  example: 456
                limit:
                  type: integer
                  description: "Number of messages to retrieve"
                  example: 20
              required:
                - conversationId
                - recipientId
                - limit
      responses:
        '200':
          description: "Messages retrieved successfully"
          content:
            application/json:
              schema:
                type: object
                description: "Response containing messages"
                properties:
                  messages:
                    type: array
                    description: "List of messages"
                    minItems: 0
                    maxItems: 1000
                    items:
                      type: object
                      description: "A message object"
                      properties:
                        messageId:
                          type: integer
                          format: int64
                          description: "ID of the message"
                          example: 789
                        senderId:
                          type: integer
                          format: int64
                          description: "ID of the sender"
                          example: 123
                        recipientId:
                          type: integer
                          format: int64
                          description: "ID of the recipient"
                          example: 456
                        content:
                          type: string
                          description: "The message content"
                          example: "Hello, how are you?"
                          minLength: 1
                          maxLength: 1000
                          pattern: "^[\\s\\S]+$"
                        timestamp:
                          type: string
                          format: date-time
                          description: "Timestamp of when the message was sent"
                          example: "2025-02-11T15:00:00Z"
                          minLength: 20
                          maxLength: 30
                          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}Z$"
                      required:
                        - messageId
                        - senderId
                        - recipientId
                        - content
                        - timestamp
                required:
                  - messages
        '400':
          description: "Bad request (invalid parameters)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Invalid request body"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
        '401':
          description: "Unauthorized (user not authenticated)"
          content:
            application/json:
              schema:
                type: object
                description: "Unauthorized error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Unauthorized access"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
...
  /message/forward:
    post:
      tags:
        - Messages
      summary: "Forward a message to other users"
      description: "Allows a user to forward a message or a photo to one or multiple users."
      operationId: forwardMessage
      security:
        - BearerAuth: []
      requestBody:
        description: "Forward message request details"
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Request object for forwarding a message"
              properties:
                text:
                  type: string
                  description: "The text of the message or the photo path if forwarding an image"
                  example: "Hello, check this out!"
                  minLength: 1
                  maxLength: 1000
                  pattern: "^[\\s\\S]+$"
                isPhoto:
                  type: boolean
                  description: "Indicates whether the forwarded content is a photo"
                  example: false
                users:
                  type: array
                  description: "List of users to whom the message is forwarded"
                  minItems: 1
                  maxItems: 1000
                  items:
                    type: object
                    description: "A recipient user object"
                    properties:
                      id:
                        type: integer
                        format: int64
                        description: "ID of the recipient user"
                        example: 12345
                    required:
                      - id
              required:
                - text
                - isPhoto
                - users
      responses:
        '200':
          description: "Message forwarded successfully"
          content:
            application/json:
              schema:
                type: object
                description: "Response indicating the message was forwarded"
                properties:
                  success:
                    type: boolean
                    description: "Indicates if the message was forwarded successfully"
                    example: true
                required:
                  - success
        '400':
          description: "Bad request (invalid request payload)"
          content:
            application/json:
              schema:
                type: object
                description: "Error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Invalid request body"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error
        '401':
          description: "Unauthorized (user not authenticated)"
          content:
            application/json:
              schema:
                type: object
                description: "Unauthorized error response"
                properties:
                  error:
                    type: string
                    description: "Error message"
                    example: "Unauthorized access"
                    minLength: 1
                    maxLength: 256
                    pattern: ".*"
                required:
                  - error